app/requirements.txt
flask==2.2.3
flask-sqlalchemy==3.0.3
flask-login==0.6.2
wtforms==3.0.1
wtforms_alchemy==0.18.0
requests==2.30.0